Streamlined interface.py
Removed all of the options stuff that was associated with outputs; it
was over-engineered and not really necessary. Flushing was starting to
make less sense on the overlays now, as the only 'flush' operation is
bliting the overlay's screens to the main display, so that's all been
moved to Output.