Substantial update to input and interface.
Now in a reasonably final form using a fully implemented system of
listeners tied to the input class. The debug listener is special and
always at the top, if it's enabled. Everything to do with debugging is
now entirely in interface/, rather than in parts of game.py

The basis of a system for creating menus has also been added, but still
needs expanding and completing.

Many other changes too:
- Updated/added docstrings
- Reorganised the interface from a file to a folder
- Moved factory.py to its own folder.
- Removed lib/getch.py as it was no longer used.
- config.py file reorganised
- Exceptions moved from strings.py to internal_strings.py
- Tidy up in several places