Now features a map editor! Can be called via 'python main.py mapeditor'
Possibly the largest single update yet. There's a lot of polish yet to
go, but this is in a workable state at the moment!

Changes:
-------

-Added a map editor, thrown together in Tkinter!

-Completely overhauled map files to use a new encoding.

-Moved factory stuff to main.py

-Moved moving vertically up and down to the basis of an 'Action' system.

-Overhauled how maps are stored internally.

-Improved falling and collision mechanics; now forming the basis of a
new system.

-Overhauled and unified ways of attaching images to class, via the new
'HasAppearances' mixin, which now gets its arguments passed via class
definition keywords rather than via a function.

-Added multiple new wall appearances. Removed default wall appearance
conversion as a result.

-Changed input handling so that multiple inputs can now be handled in a
single tick. Not that likely to be a big concern but nice to have.

-Enhanced pygame.Surfaces to use new offset stuff; improved cutouts.
Added new methods taking advantage of the new offset stuff.

-Fixed a bug in which clicking on entries twice made them appear to be
deselected.

-Reworked tile system slightly to fit in with the new appearane system,
and with the map editor.