Several interface improvements.
Firstly, have reworked the interface system to bring together overlays
and listeners. This has allowed for a re-tidy of the various components
forming an interface.

Secondly, menus now implement event-driven callbacks, which now allows
for more general menus, see the smoother implementation of the map
select menu. See in particular how we can now generate a messagebox on
the fly.

Also:
-Attempting to load a corrupted file now displays an error message box.
-Console lines now wrap
-Can now create messageboxes