Added an 'escape' menu.
The 'escape' menu appears when pressing escape in the main game, and
gives options to quit the game.

Also switched many uses of tools.Object to the more efficient
collections.namedtuple, in those cases where mutability is not required.

Also reworked how the tiles are yielded in game.Map.local, so that they
are precisely the ones which collide with the given disc, and they are
yielded going outwards from the center.

Improved how the selectd overlay works, so that it always gives the most
recently selected currently-enabled overlay.

Added an interface (technical term) to the interface (non technical
term. Hmm, that really needs renaming already.) for overlays to use.

Hooked sdl.Surface.get_parent and get_abs_parent into using a _parent
attribute that is set when applying cutouts, so that cutouts act a bit
better, as compared to subsurfaces. Reworked get_abs_offset to also take
advantage of this.